+++ static/__tests__/dtale/DataViewer-reload-test.jsx
@@ -48,7 +48,7 @@ describe("DataViewer tests", () => {
     Object.defineProperty(HTMLElement.prototype, "offsetWidth", originalOffsetWidth);
   });
 
-  test("DataViewer: base operations (column selection, locking, sorting, moving to front, histograms,...", done => {
+  test("DataViewer: base operations (column selection, locking, sorting, moving to front, col-analysis,...", done => {
     const { DataViewer, ReactDataViewer } = require("../../dtale/DataViewer");
 
     const store = reduxUtils.createDtaleStore();