+++ static/main.jsx
@@ -8,9 +8,9 @@ import "./adapter-for-react-16";
 import { DataViewer } from "./dtale/DataViewer";
 import { CodeExport } from "./popups/CodeExport";
 import { CodePopup } from "./popups/CodePopup";
+import { ReactColumnAnalysis as ColumnAnalysis } from "./popups/ColumnAnalysis";
 import { ReactCorrelations as Correlations } from "./popups/Correlations";
 import { ReactDescribe as Describe } from "./popups/Describe";
-import { ReactHistogram as Histogram } from "./popups/Histogram";
 import Instances from "./popups/Instances";
 import { ReactCharts as Charts } from "./popups/charts/Charts";
 import { ReactCreateColumn as CreateColumn } from "./popups/create/CreateColumn";
@@ -39,8 +39,8 @@ if (_.startsWith(window.location.pathname, "/dtale/popup")) {
     case "build":
       rootNode = <CreateColumn {...{ dataId, chartData }} />;
       break;
-    case "histogram":
-      rootNode = <Histogram {...{ dataId, chartData }} height={250} />;
+    case "column-analysis":
+      rootNode = <ColumnAnalysis {...{ dataId, chartData }} height={250} />;
       break;
     case "instances":
       rootNode = <Instances dataId={dataId} iframe={true} />;
@@ -56,7 +56,7 @@ if (_.startsWith(window.location.pathname, "/dtale/popup")) {
   ReactDOM.render(rootNode, document.getElementById("popup-content"));
 } else if (_.startsWith(window.location.pathname, "/dtale/code-popup")) {
   require("./dtale/DataViewer.css");
-  document.getElementById("code-title").innerHTML = `${window.opener.code_popup.title} Code Snippet`;
+  document.getElementById("code-title").innerHTML = `${window.opener.code_popup.title} Code Export`;
   ReactDOM.render(<CodePopup code={window.opener.code_popup.code} />, document.getElementById("popup-content"));
 } else {
   const store = createStore(app.store);